update:删除inst 文件，统一用output.go
